requirements.txt
python-dotenv
pandas
numpy
seaborn
matplotlib
flask
pymongo
certifi
scikit-learn
pyaml
# -e .